add registry-redis unit
