+++ spi-samples-configcenter/spi-samples-configcenter-etcd/src/main/java/org/apache/dubbo/spi/samples/configcenter/EtcdConsumer.java
@@ -17,10 +17,22 @@
  *
  */
 
-package org.apache.dubbo.spi.samples.registry.api;
+package org.apache.dubbo.spi.samples.configcenter;
 
-public interface DemoService {
 
-    String sayHello(String name);
+import org.apache.dubbo.spi.samples.configcenter.api.DemoService;
 
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+
+public class EtcdConsumer {
+
+    public static void main(String[] args) {
+        ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(new String[]{"spring/configcenter-consumer.xml"});
+        context.start();
+        DemoService demoService = context.getBean("demoService", DemoService.class);
+
+        String hello = demoService.sayHello("world");
+        System.out.println(hello);
+
+    }
 }
