+++ spi-samples-configcenter/spi-samples-configcenter-etcd/src/test/java/registry/DemoServiceIT.java
@@ -18,7 +18,7 @@
 package registry;
 
 
-import org.apache.dubbo.spi.samples.registry.api.DemoService;
+import org.apache.dubbo.spi.samples.configcenter.api.DemoService;
 
 import org.junit.Assert;
 import org.junit.Test;
